add a pulse for whatsapp accounts #359
This adds an interval that sends presence requests for every chat every 30 seconds to keep the connection to whatsapp alive.
I tested it on my Geeksphone Peak and I can say that I wasn't disconnected the entire day. I was most of the time on wifi and the battery usage didn't increased, but I'll do future tests on mobile data, just to make sure.