WhatsApp pulse ratelimit hot fix
I had to discover that whatsapp quickly starts to block the status request overflow with an error message. The server only responses with the requested data again, when we stop our request for some itme.
Because of this, I dissabled the request of the statuses and the pulse will now just ask for the pictures. WhatsApp does not block these requests.