better and shorter pulse
after the feedback from @cmeerw I took an additional look at the CoSeMe library and I found the pong feature to keep the stream alive. I changed the pulse to send a pong. In addition I moved the pulse from the account into the CoSeMe connector. I think this is better now.