Stacktrace and Logfiles
- it's easier to locate an log message from `Tools.log()` if we have a stacktracke
- now we can write our logs onto the SD card, this helps us to get logs from users who don't want to mess with WebIDE

The file logger is pretty slow and I think it uses a lot of CPU. I also can't tell if it will cause a crash on Devices with less memory than the Flame.
Both features need to be enabled from the dev setings, but because I already messed with app.js for #861 I will land these new settings with the next patch.